Alertify is now a full fledged Go package
Turning the project into a package allows to reuse it in other programs.
It now allows to register various monitoring sources with the main bot
as long as they implement particular interface.

Slack parts have been ripped out of the core package and moved to a
dedicated monitor package.

The original full program has been moved to examples directory and is
now called `slackertify`.